"double" hp
more complex encounter

+++ Entity/Player/Player.hpp
@@ -2,7 +2,7 @@
 
 #include "../Entity.hpp"
 #include "../../Inventory.hpp"
-
+#include "../../Item/Item.h"
 
 
 class Player : public Entity
@@ -14,7 +14,7 @@ class Player : public Entity
 
     public:
         Player();
-        Player(int h, int wl, int fl);
+        Player(double h, int wl, int fl);
         
         void adventure();
